refactor: make product home page a web extension
